requirements.txt
corner==2.2.1
keras==2.11.0
matplotlib==3.6.2
numpy==1.24.2
scikit-learn==1.2.1
scipy==1.9.3
uproot==5.0.2

